Define an "api client", which is just an axios client with no additional configuration. However, we can use this as an entrypoint to centralize behaviors of the API client. For one example, setting the base url for requests.
The step taken here is to create the client file, and change any api-interaction services to use the client rather than referencing axios directly.

Additionally, this commit defines a logger service module, which proxies to the methods of `console`, but has utilities for dynamically adjusting the log level. This may be useful in the future, for e.g. minimizing logging during testing, or supplying debug logs for common trouble spots, or trace logs for areas known to be sensitive to performance.

Signed-off-by: Neal Ormsbee <neal.ormsbee@ibm.com>