This commit does two things: (#119)
1. Add the Prettier code formatter to package.json as a devDependency.
2. Execute `npx prettier -w ./src`, which has the effect of formatting all source files using the default Prettier config.

I would like to make code-formatting using this tool a requirement for contribution, but imagine that a well-formed CONTRIBUTING.md should be added with basic requirements beyond formatting if we're going down that road.
I'd like to take some time to think about that document and put forth a real effort for the community's consideration, and so it has been omitted for now.

Signed-off-by: Neal Ormsbee <neal.ormsbee@ibm.com>